Adds deeplink handler if it is found
